requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
backports-zoneinfo==0.2.1
    # via django
black==22.3.0
    # via -r requirements.in
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via black
django==4.0.5
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
idna==3.3
    # via requests
mccabe==0.6.1
    # via flake8
mypy==0.960
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
requests==2.27.1
    # via -r requirements.in
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   black
    #   mypy
types-requests==2.27.30
    # via -r requirements.in
types-urllib3==1.26.15
    # via types-requests
typing-extensions==4.2.0
    # via
    #   black
    #   mypy
urllib3==1.26.9
    # via requests
